requirements.txt
python-dotenv>=1.0.0
google-cloud-bigquery>=3.17.0
google-cloud-storage>=2.17.0
google-cloud-logging>=3.9.0
langchain>=0.1.16
langchain-anthropic>=0.1.0
langchain-community>=0.0.38
langgraph>=0.0.53
langchain-core>=0.1.52
chromadb>=0.4.24
sentence-transformers>=2.2.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.18.0
streamlit>=1.31.0
requests>=2.31.0
pydantic>=2.7.0
db-dtypes>=1.0.0
